requirements.txt
os
pillow
json
